whtapi/requirements.txt
Flask==3.0.2
flask-socketio==5.3.6
python-engineio==4.8.2
python-socketio==5.11.1
gevent==23.9.1
gevent-websocket==0.10.1
pymongo==4.6.2
pyotp==2.9.0
PyJWT==2.8.0
python-dotenv==1.0.1
flask-cors==4.0.0 